refactor(rpc): rename method `get_transactions` to `get_ledger_transactions`
